Add AUTHORIZED_KEYS to manage public keys file
This should be set in local.conf, if this variable points to invalid file this considerate as an fatal error

+++ recipes-connectivity/openssh/openssh_8.0p1.bbappend
@@ -1,15 +1,21 @@
 FILESEXTRAPATHS_prepend := "${THISDIR}/files:"
 
-SRC_URI += "file://authorized_keys \
-            file://sshd_config \
+AUTHORIZED_KEYS ?= ""
+SRC_URI += "file://sshd_config \
             "
 
 FILES_${PN} += " \
     ${ROOT_HOME}/.ssh \
-    ${ROOT_HOME}/.ssh/authorized_keys \
 "
 
 do_install_append() {
-    install -d ${D}${ROOT_HOME}/.ssh
-    install -m 0644 ${WORKDIR}/authorized_keys ${D}${ROOT_HOME}/.ssh
+	bbnote "AUTHORIZED_KEYS is set to invalid value : \"${AUTHORIZED_KEYS} \" "
+
+	if [ -e "${AUTHORIZED_KEYS}" ]; then 
+		install -d ${D}${ROOT_HOME}/.ssh
+		install -m 0644 ${AUTHORIZED_KEYS} ${D}${ROOT_HOME}/.ssh/authorized_keys
+	else
+		bbfatal "ERROR: AUTHORIZED_KEYS variable is set to invalid value file does not exist : \"${AUTHORIZED_KEYS} \". Change your local.conf "
+	fi
+	
 }
